pyproject.toml
[build-system]
requires = [
  "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "dbnet"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "gurimusan", email = "gurimusan@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "torch",
  "torchvision",
  "torchinfo",
  "dataclass-wizard",
  "tqdm",
  "numpy",
  "pillow",
  "requests",
  "scipy",
  "opencv-python",
  "albumentations",
  "pyclipper",
  "shapely",
]

[project.optional-dependencies]
test = [
    "pytest-mock",
]

[project.urls]
Documentation = "https://github.com/gurimusan/dbnet#readme"
Issues = "https://github.com/gurimusan/dbnet/issues"
Source = "https://github.com/gurimusan/dbnet"

[tool.hatch.version]
path = "src/dbnet/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dbnet tests}"

[tool.coverage.run]
source_pkgs = ["dbnet", "tests"]
branch = true
parallel = true
omit = [
  "src/dbnet/__about__.py",
]

[tool.coverage.paths]
dbnet = ["src/dbnet", "*/dbnet/src/dbnet"]
tests = ["tests", "*/dbnet/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
